chore: split embed module
